@tqchen 2018-02-13
@merrymercy do you think if it is possible to generalize to batch_size > 1?