@tqchen 2018-02-13
It is of interest to bring it to other backends as well, so I opened an issue here https://github.com/dmlc/tvm/issues/899