@merrymercy 2018-02-14
@tqchen It depends on the GEMM implementation. Now It works well for batch size 1~64 without loss of peak performance.